requirements.txt
torch==1.7.1 
torchvision==0.8.2 
numpy==1.19.5
pandas==1.1.5
regex==2019.12.20
scikit-image==0.16.2
scikit-learn==0.24
scipy==1.4.1
seaborn==0.11.1
seqeval==1.2.2
sklearn-pandas==1.8.0
tqdm==4.41.1
torchtext==0.9.1
pytorch-fid==0.2.0
pytorch-msssim==0.2.1